app/requirements.txt
amqp==5.0.5
asgiref==3.3.1
backcall==0.2.0
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
decorator==4.4.2
Django==3.1.5
django-annoying==0.10.6
django-crispy-forms==1.11.1
django-debug-toolbar==3.2
django-extensions==3.1.1
Faker==6.4.1
flake8==3.8.4
flake8-builtins==1.5.3
flake8-print==4.0.0
importlib-metadata==3.4.0
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.0.2
mccabe==0.6.1
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.17
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.8.1
python-dateutil==2.8.1
pytz==2020.5
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==4.3.3
typing-extensions==3.7.4.3
vine==5.0.0
wcwidth==0.2.5
xlwt==1.3.0
zipp==3.4.0
